added kinesis and firehose examples in boto3
